update default symbol table and apply offset on block symbols
The change comes with a small refactoring that attempts to make things a
bit more typesafe by encapsulating symbol-table-related operations in a
dedicated module. The result is not entirely satisfying but a step in
the right direction.

+++ biscuit/src/Auth/Biscuit/Proto.hs
@@ -113,7 +113,7 @@ data PredicateV2 = PredicateV2
     deriving anyclass (Decode, Encode)
 
 data TermV2 =
-    TermVariable (Required 1 (Value Int32))
+    TermVariable (Required 1 (Value Int64))
   | TermInteger  (Required 2 (Value Int64))
   | TermString   (Required 3 (Value Int64))
   | TermDate     (Required 4 (Value Int64))
